requirements.txt
PyYAML
Flask
flask-sqlalchemy
flask-bcrypt
sqlalchemy
werkzeug
pymysql
cryptography